Moved FitResults to a seperate file
